try to evaluate

+++ src/models/io.py
@@ -64,7 +64,7 @@ def upsample_wav(wav, model):
 	outname = wav + '.' + out_label
 	rand = 4
 	librosa.output.write_wav(outname + '.hr.wav', x_hr, fs)	
-	librosa.output.write_wav(outname + '.lr.wav', (int)x_lr, fs / int(rand))	
+	librosa.output.write_wav(outname + '.lr.wav', int(x_lr), fs / rand)	
 	librosa.output.write_wav(outname + '.pr.wav', x_pr, fs)	
 
 	# save the spectrum
